package.json: Import libp2p and not separate deps
This project builds on top of libp2p but imports separate
libp2p libraries instead of the main package. This makes it
difficult to use while using libp2p. i.e. when a host has been
created with libp2p.New() and the libp2p-peer version does not
match the one here.