Fix simultaneous connection bug
When initializing a PubSub connection to another peer, we now inspect
the peer ID of the peer that initiated a `Conn`. When encountering
subsequent connections to a peer, we measure (in integral space) the
initiating peer ID of THAT connection against the existing connection.
If, and only if, that peer ID is GREATER THAN the existing initiator
peer ID, we close our old stream and open a stream on that new
connection. This unambiguous logic ensures that both peers maintain the
same stream, instead of closing their stream (the first to notify in a
simultaneous connect scenario) and favoring the other peer's stream,
which the other peer will in turn close.