stop timer in a way that works befor and afgter go1.23
In this case, a stopped timer is created. Since the timer duration is set to 1 hour and is immediately stopped, the timer shoud never fire and the channel should never need to be cleared (as is generally needed for go1.22). For go-1.23 the timer semantics have changed so that reading the timer channle following Stop will always block. Therefore, for both these casses, only calling Stop is sufficient to crete the stopped timer.