partialmessages: add explicit MergePartsMetadata function
This allows a peer's local partsMetadata to merge with their remote
partsMetadata. As long as the underlying PartsMetadata type supports
CRDT-like addition, we can avoid sending the peer duplicates.

Take the following example (credit Sukun):

> Say: We have parts 1, 2, 5, 6
> 
> 1. Peer sends a message saying I have part 1, 2
> 2. We send parts 5, 6 and update the peerstate in PublishMessage to (1, 2, 5, 6)
> 3. Concurrently with 2, we receive message from the peer saying I now have parts (1, 2, 3, 4)

With this change we can now correctly store the fact that this peer
should have parts 1,2,3,4,5,6. And in a way that does not assume a
specific representation of the PartsMetadata.