ensure that the Hello Packet is the first rpc sent
When a topic join happens concurrently with adding a new peer, we may
send the topic join announcement before the hello packet which contains
our extensions. This is a protocol violation.

This change starts sending rpcs on new stream only after we've enqued a
hello packet. We use UrgentPush to ensure that it's the first one sent.